[refactor] SearchViewModel ChallegneFetchUsecase, ImageFetchUsecase 적용
