+++ Routinus/Routinus/Application/Coordinator/SearchCoordinator.swift
@@ -21,8 +21,11 @@ final class SearchCoordinator: RoutinusCoordinator {
 
     func start() {
         let repository = RoutinusRepository()
-        let searchUsecase = SearchFetchUsecase(repository: repository)
-        let searchViewModel = SearchViewModel(category: category, usecase: searchUsecase)
+        let imageFetchUseccase = ImageFetchUsecase(repository: repository)
+        let challengeFetchUsecase = ChallengeFetchUsecase(repository: repository)
+        let searchViewModel = SearchViewModel(category: category,
+                                              imageFetchusecase: imageFetchUseccase,
+                                              challengeFetchUsecase: challengeFetchUsecase)
         let searchViewController = SearchViewController(with: searchViewModel)
         self.navigationController.pushViewController(searchViewController, animated: true)
